requirements.txt
torch
pillow
pandas
numpy
albumentations
torchvision
matplotlib
torchinfo
torchmetrics
segmentation-models-pytorch
